Quote api create done and setup

+++ server/controllers/QuotationController.js
@@ -1,86 +1,68 @@
-const Quote = require('../models/QuoteModel');
+const { handleApiResponse } = require('../modules/responseHandler');
 const { ViewQuote, AddQuote, SingleQuote, DeleteQuote, UpdateQuote } = require('../services/QuoteService');
+const { idSchema } = require('../validators/Schemas');
 
-const { calculate } = require('../helpers/helpers');
-
+// To get All Quotes List
 const getAllQuotes = async (req, res) => {
     try {
-        const { sort, select, page, limit } = req.query;
-
-        const Quotes = await ViewQuote({
-            sort,
-            select,
-            page: Number(page) || 1,
-            limit: Number(limit) || 10,
-        });
+        const Quotes = await ViewQuote(req.query);
 
         if (!Quotes || Quotes.length === 0) {
-            return res.status(404).json({
-                Status: 'success',
-                Message: 'Quotes Not found',
-                Quotes: [],
-                nbHits: 0,
-            });
+            return handleApiResponse(res, 404, 'Quotes not found');
         }
 
-        res.status(200).json({
-            Status: 'success',
-            Message: 'Quotes fetched successfully',
-            Quotes,
+        handleApiResponse(res, 200, 'Quotes  fetched successfully', {
+            data: Quotes,
             nbHits: Quotes.length,
         });
     } catch (error) {
-        res.status(500).json({
-            message: 'An error occurred while fetching Quotes',
-            error: error.message,
-        });
+        handleApiResponse(res, 500, 'An error occurred while fetching the Quotes', { error: error.message });
     }
 };
-
-const createQuote = async (req, res) => {
-    const data = req.body;
-    try {
-        const savedQuote = await AddQuote(data);
-        res.status(201).json({
-            status: 'success',
-            message: 'Quote added successfully',
-            savedQuote,
-        });
-    } catch (error) {
-        if (error.message.includes('Quote with this Quote No. already exists')){
-            res.status(400).json({
-                status: 'fail',
-                error: `${error.message}`,
-                message: 'Quote with this Quote No. already exists',
-            });
-        } else {
-            res.status(500).json({
-                status: 'error',
-                message: `An error occurred while adding the Quote: ${error.message}`,
-            });
-        }
-    }
-};
-
 // To get Single Quote Details
 const getSingleQuote = async (req, res) => {
     try {
         const id = req.params.id;
+        await idSchema.parseAsync({ _id: id });
         const Quote = await SingleQuote(id);
 
         if (!Quote) {
-            return res.status(404).json({ message: 'Quote not found' });
+            return handleApiResponse(res, 404, 'Quote not found');
         }
-        res.status(200).json({
-            status: 'success',
-            message: 'Quote details fetched successfully',
-            Quote: Quote,
+
+        handleApiResponse(res, 200, 'Quote  details fetched successfully', {
+            data: Quote,
+            nbHits: 1,
         });
     } catch (error) {
-        res.status(500).json({
-            message: 'An error occurred while fetching the single Quote',
-            error: error.message,
+        const errorMessage = error.message.includes('Invalid ID format') ? 'Use a Proper Id' : `An error occurred while fetching the Quote: ${error.message}`;
+        handleApiResponse(res, error.message.includes('Invalid ID format') ? 400 : 500, errorMessage, { error: 'An error occurred while fetching the Quote' });
+    }
+};
+
+// To Add a Quote to Quotes list
+const postSingleQuote = async (req, res) => {
+    try {
+        const Quote = await AddQuote(req.body);
+
+        handleApiResponse(res, 201, 'Quote added successfully', {
+            data: Quote,
         });
+    } catch (error) {
+        const duplicateFieldMatches = error.message.match(/[a-zA-Z_]+(?= already exists)/g);
+        if (duplicateFieldMatches && duplicateFieldMatches.length > 0) {
+            const duplicateFields = duplicateFieldMatches.map((field) => field.replace('quote_', ''));
+            const errorMessage = `Quote with ${duplicateFields.join(', ')} is already exists.`;
+            handleApiResponse(res, 400, errorMessage, error);
+        } else if (error.message.includes('Client not found')) {
+            handleApiResponse(res, 404, 'Client not found', { error: error.message });
+        } else if (error.message.includes('User not found')) {
+            handleApiResponse(res, 404, 'User not found', { error: error.message });
+        } else if (error.message.includes('Product not found')) {
+            handleApiResponse(res, 404, 'Product not found', { error: error.message });
+        } else {
+            handleApiResponse(res, error.status || 500, error.message || 'Internal server error');
+        }
     }
 };
 
@@ -88,22 +70,22 @@ const getSingleQuote = async (req, res) => {
 const deleteSingleQuote = async (req, res) => {
     try {
         const id = req.params.id;
-        const Quote = await DeleteQuote(id);
+        await idSchema.parseAsync({ _id: id });
 
-        if (!Quote || Quote.deletedCount === 0) {
-            return res.status(404).json({ message: 'Quote not found, deletion unsuccessful' });
+        const DeletedQuote = await SingleQuote(id);
+
+        if (!DeletedQuote) {
+            return handleApiResponse(res, 404, 'Quote not found, deletion unsuccessful');
         }
 
-        res.status(200).json({
-            status: 'success',
-            message: 'Quote deleted successfully',
-            deletedQuote: Quote,
+        const DeletedQuoteRes = await DeleteQuote(id);
+
+        handleApiResponse(res, 200, 'Quote deleted successfully', {
+            deleted: DeletedQuoteRes,
         });
     } catch (error) {
-        res.status(500).json({
-            message: 'An error occurred while deleting the single Quote',
-            error: error.message,
-        });
+        const errorMessage = error.message.includes('Invalid ID format') ? 'Use a Proper Id' : `An error occurred while deleting the Quote: ${error.message}`;
+        handleApiResponse(res, error.message.includes('Invalid ID format') ? 400 : 500, errorMessage, { error: 'Internal Server Error' });
     }
 };
 
@@ -111,6 +93,7 @@ const deleteSingleQuote = async (req, res) => {
 const updateSingleQuote = async (req, res) => {
     try {
         const id = req.params.id;
+        await idSchema.parseAsync({ _id: id });
         const updateQuoteData = req.body;
 
         const updatedQuote = await UpdateQuote(id, updateQuoteData);
@@ -119,16 +102,27 @@ const updateSingleQuote = async (req, res) => {
             return res.status(404).json({ message: 'Quote not found, update unsuccessful' });
         }
 
-        res.status(200).json({
-            status: 'success',
-            message: 'Quote updated successfully',
-            updatedQuote,
+        handleApiResponse(res, 200, 'Quote updated successfully', {
+            data: updatedQuote,
         });
     } catch (error) {
-        res.status(500).json({
-            message: 'An error occurred while updating the single Quote',
-            error: error.message,
-        });
+        const errorMessage = error.message.includes('Invalid ID format') ? 'Provide valid Id' : `An error occurred while updating the single Quote: ${error.message}`;
+        if (errorMessage === 'Provide valid Id') {
+            handleApiResponse(res, 400, errorMessage, { error: 'Internal Server Error' });
+        } else {
+            if (error.message.includes('E11000 duplicate key error')) {
+                handleApiResponse(res, 500, 'Quote Name must be unique', { error: error.message });
+            } else {
+                handleApiResponse(res, 500, errorMessage, { error: 'Internal Server Error' });
+            }
+        }
     }
 };
-module.exports = { getAllQuotes, createQuote, getSingleQuote, deleteSingleQuote, updateSingleQuote };
+
+module.exports = {
+    getAllQuotes,
+    getSingleQuote,
+    postSingleQuote,
+    deleteSingleQuote,
+    updateSingleQuote,
+};
